RM#54967 PO Line and Invoice Line : set translation only if there is no supplier catalog product
