+++ changelogs/unreleased/54967.yml
@@ -1,0 +1,3 @@
+---
+title: "Purchase Order Line : now correctly takes the product name of the supplier catalog if there is one."
+type: fix
